+++ valkey_z_php_methods.c
@@ -822,6 +822,10 @@ BGSAVE_METHOD_IMPL(ValkeyGlide)
 BGREWRITEAOF_METHOD_IMPL(ValkeyGlide)
 /* }}} */
 
+/* {{{ proto bool ValkeyGlide::reset() */
+RESET_METHOD_IMPL(ValkeyGlide)
+/* }}} */
+
 /* {{{ proto bool ValkeyGlide::save() */
 SAVE_METHOD_IMPL(ValkeyGlide)
 /* }}} */